Fix creep_utils.count_roles() to take into account different room contexts.
TODO: Make role_count a per-room variable

This also changes spawning code to not try and spawn a 1-work harvester as soon as possible if our only creep is a 5-work harvester. Now, instead of having the emergency spawning code hardcode to spawning 1-work creeps, it will wait until there's enough energy for at least a creep a bit more than 1/2 as large as the work_mass present in the room!