Define a new creep_utils function, distance_squared_room_pos, which gets the squared distance between two RoomPositions, **taking into account the room distance**. Use this in all places in TargetMind where getRange was used.
This is a replacement for RoomPosition.getRangeTo(RoomPosition, int), which A. is less efficient due to trying to find exact range rather than squared range, and B. returns Infinity when the two positions are in different rooms, rather than parsing the room names and calculating distance based on 50 block wide rooms!

This commit also adds range checking to the TargetMind _find functions for remote miners, and remote haulers - to prepare for a future where multiple rooms might be fully owned!