Fix remote mining flags, and hopefully get remote mining working for multiple rooms, each with indipendent targets.
The "sponsor" of a remote mine, the one mining it, is set via the "flag.memory.sponsor" value, or Memory.flags.flag_name.sponsor in memory.

This also significantly lowers the minimum work mass for starting a remote mining operation, in an attempt to get our first colony (with only one energy source) also remote mining.