Attempt to change all spawning-related code for workers and haulers to create dynamic creep bodies.
Now, for all roles with med_hauler or worker base, the number of WORK or CARRY parts to spawn is calculated instead of the number of creeps to spawn. We might now actually spawn too many, since there is no spawning fewer-section creeps if we're only a few work or carry short, but it should work well besides that.

This is in attempt to fix a CPU shortage caused by too many creeps - we also weren't using our full spawn capacity at all before.