Add a new generic role, Recycling.
For mostly all creeps which already committed suicide, recycling is now used instead. In addition, Replacing creeps make their Replaced creeps recycling, and Builders, Local Haulers, Spawn Fillers, and Tower Fillers now automatically transition to Recycling when they have less than 50 ticks of life left.

A recycling creep will first check if it has any resources of any type in its carry, and if it does it will go to storage and offload them. After doing that, it will go to spawn, and will recycling itself.

This is primarily useful for spawnfill, local haulers and tower fillers with the new larger creep size, so that if a creep with say, 300 capacity has been waiting at a dedicated miner and it's 75% full, but will die before it fills all the way, it will actually get the energy to the storage before it dies!