Implement "work stealing" for TargetMind, where a creep will "steal" a target from another creep if the closest target is at least 1.5 closer to him than the previously targeting creep.
This is only enabled for spawn fillers for now, but may be useful for other roles in the future!

This greatly improves the efficiency of spawn fillers when some of them are out providing energy to worker creeps while spawn is inactive, as now the closest creep to spawn can go and fill as many extensions as it can, and not wait for far away creeps who've reserved the right to fill certain extensions or the spawn.