Overhaul MineralHauler to have 5 different states:
Now, when MineralHauler doesn't have minerals to gather and is trying to fill the terminal with energy, it will switch to a "terminal emptying" state where it takes minerals out of the terminal and puts them into storage until the terminal is less than 75% full.

This is very useful when gathering a lot of energy to send in the terminal (via the memory setting), and there's not enough room in it. It all works now!