Improve HoneyTrails... maybe
This essentially completely duplicates the cost matrix generation code in HoneyTrails, for the sake of allowing serialization and use of serialized cost matrices for in-between rooms for remote mining and pathing.

It works, but it's quite ugly and will probably want to be improved in the future.

This also adds a "max avoidance" parameter to find_path (and like methods) which allows military creeps for example to highly avoid an enemy room. This is already in use in the new TowerDrainer, which shouldn't have been committed as part of the defense commit, but was anyways.