Add support for rerouting some movement through a specific portal.
This is currently very hardcoded - a flag needs to be placed at the origin portal named 'reroute' and another at the destination portal named 'reroute_destination'.

The code does no checking of whether these flags are actually at a portal, it just blindly combines the origin->reroute and reroute_destination->destination paths in pathfinder. It does do checks to ensure that the portal is only used if the linear range from origin->portal + portal_dest->destination is smaller than origin->destination, so no need to manually reroute creeps besides just placing the flag.

In the future, multiple such flags should be allowed, and they should be auto-placed when a portal is detected and auto-removed when the portal is dead. One other nice feature would be to auto-turn-on prio_spawn for any colonies which we can only reach through the portal when the portal starts to decay.