Add a new room balancing feature!
This now makes rooms balance depending, not on storage contents, but on wall height! The RCL8 room with the highest walls will now send energy to the RCL6+ room with the lowest walls.

Energy is sent via a fulfillment order, and it will calculate how much energy can be sent while keeping at least both 300,000 total resources in the room and 100,000 energy!

Wall height in this commit refers not to the biggest wall, but the smallest wall in each room. So the RCL8 room in which it's smallest wall is bigger than all other RCL8 rooms smallest walls will be the one sending energy.