Improve miner movement, so it will move directly to it's prefered position if it already knows where one is.
This prevents miners from doing a back-and-forth movement if the path from one position next to the mine to the prefered position is more than one block long (i.e. if they have to go around something). Beforehand, logic would switch to moving directly to the mine as soon as they stepped away, so they would never fully mine out the source, and would just move back and forth forever.