Huge overhall of road placement, for both speed and improved road reuse!
This should help a ton, and although there are a few bugs still with recording exactly where the roads are, it should be an improvement.

To clarify: the 'extraneous road cleanup' part of the code is currently disabled, as something causes it to want to delete some (but not all) of existing legitimate road construction sites.