Make haulers sit between terminal/storage if possible!
This adds two utility functions: one which checks for available positions between two positions in the same room, serializes the results and returns them, and another which finds the closest of multiple serialized positions to a non-serialized position in the same room.

Together, these are used to cache the open spaces between the storage and the terminal, and to have mineral haulers move towards them rather than terminal or storage when picking up minerals. (behavior when dropping off minerals is still the same, since it's less likely they will head to the other of terminal/storage if they are dropping off, and if they do they'll be picking up and head to the shared spot)