Fix kiting attackers!
A while ago, in the major refactor, it seems a bug was introduced into kiting attackers which caused an infinite loop in cost matrix generation when running away.

This was due to Transcrypt turning `for x in range(x - drange, x + drange + 1)` into `for (var x = x - drange; x < x + drange + 1; x++)` (note the lack of a temporary variable like python would have for 'range').

In any case, that bug was fixed, but I also rewrote the helper function in question to perform more efficiently/effectively (hopefully!)