Remove checking memory before loading all code.
I've checked to ensure this, and it seems indeed that no code uses Memory now outside of the main loop. This allows us to not ever check memory for parsing time outside of the main loop, and to clean up main.py a bit.

This also cleans up main.py's constants imports.