Move all constants from screeps_constants to a module named screeps.
This also includes the actual values/types for all API constants, and allows for the framework of including future types as well in separate files!

This works well with PyCharm, I haven't tested with any other editors.

In addition to the editor/autocomplete support this allows, it also helps by separating out new_set and new_map functions, so we don't actually need to be importing them _everywhere_